+++ engine/README.md
@@ -1,3 +1,3 @@
-# @dpkit/service
+# @dpkit/engine
 
 dpkit is a fast data management framework built on top of the Data Package standard and Polars DataFrames. It supports various formats like CSV, JSON, and Parquet and integrates with data platforms such as CKAN, Zenodo, and GitHub. For more information, please visit the [project's website](https://dpkit.app).
