+++ toolbox/package.json
@@ -25,7 +25,7 @@
     "build": "tsc"
   },
   "dependencies": {
-    "@dpkit/service": "workspace:*"
+    "@dpkit/engine": "workspace:*"
   },
   "devDependencies": {
     "react-strict-dom": "0.0.54"
