+++ website/services/api.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@dpkit/service/client"
+import { createClient } from "@dpkit/engine/client"
 
 export const api = createClient({
   url: import.meta.env.VITE_API_URL,
